Showing XP in the HUD

+++ Source/Aura/Private/UI/WidgetController/OverlayWidgetController.cpp
@@ -117,7 +117,7 @@ void UOverlayWidgetController::OnXPChanged(int32 NewXP) const
 		const int32 DeltaLevelRequirement = LevelUpRequirement - PreviousLevelUpRequirement;
 		const int32 XPForThisLevel = NewXP - PreviousLevelUpRequirement;
 		
-		const float XPBarPercent = static_cast<float>(XPForThisLevel / DeltaLevelRequirement);
+		const float XPBarPercent = static_cast<float>(XPForThisLevel) / static_cast<float>(DeltaLevelRequirement);
 		
 		OnXPPercentChangeDelegate.Broadcast(XPBarPercent);
 	}
